remove rdb repo
